Fix `COMMAND_CLEAR_CAR_LAST_WEAPON_DAMAGE`

+++ source/game_sa/Scripts/Commands/VehicleCommands.cpp
@@ -919,7 +919,7 @@ eModelID GetCarModel(CVehicle& self) {
 * @param self CVehicle&
 */
 void ClearCarLastWeaponDamage(CVehicle& self) {
-    self.m_nLastWeaponDamageType = WEAPON_NONE;
+    self.m_nLastWeaponDamageType = (uint8)(WEAPON_NONE);
     assert(self.m_nLastWeaponDamageType == (uint8)(-1)); // sanity check
 }
